wallet: move light wallet RPC out of core RPC
It's not nothing to do with it
